hud control framework
