+++ addons/common/XEH_PREP.hpp
@@ -105,6 +105,7 @@ PREP(handleModifierKeyUp);
 PREP(hasItem);
 PREP(hasMagazine);
 PREP(headBugFix);
+PREP(hideHUDHelper);
 PREP(hideUnit);
 PREP(interpolateFromArray);
 PREP(inTransitionAnim);
@@ -139,8 +140,10 @@ PREP(positionToASL);
 PREP(progressBar);
 PREP(readSettingFromModule);
 PREP(receiveRequest);
+PREP(registerHUDElement);
 PREP(registerItemReplacement);
 PREP(removeCanInteractWithCondition);
+PREP(removeHUDElement);
 PREP(removeSpecificMagazine);
 PREP(requestCallback);
 PREP(resetAllDefaults);
