+++ addons/medical_feedback/XEH_PREP.hpp
@@ -1,4 +1,3 @@
-PREP(createHUDIndicators);
 PREP(effectBleeding);
 PREP(effectBloodVolume);
 PREP(effectBloodVolumeIcon);
