--- addons/medical_feedback/functions/fnc_createHUDIndicators.sqf
@@ -1,18 +1,0 @@
-#include "script_component.hpp"
-/*
- * Author: BaerMitUmlaut, modified by GhostIsSpooky
- * Creates the HUD indicators.
- *
- * Arguments:
- * None
- *
- * Return Value:
- * None
- *
- * Example:
- * call ace_medical_feedback_fnc_createHUDIndicators
- *
- * Public: No
- */
-
-QGVAR(HUDIndicators) cutRsc [QGVAR(HUDIndicatorsDisplay), "PLAIN", 0.01, false];
