+++ addons/ui/RscTitles.hpp
@@ -5,20 +5,12 @@
 class RscPictureKeepAspect;
 
 class RscTitles {
-    class GVAR(speedIndicatorDisplay) {
-        idd = -1;
-        movingEnable = "true";
-        enableSimulation = 1;
-        enableDisplay = 1;
-        duration = 999999;
-        fadeIn = 0;
-        fadeOut = 0;
-        name = QGVAR(speedIndicatorDisplay);
-        onLoad = QUOTE(uiNamespace setVariable [ARR_2(QQGVAR(speedIndicatorDisplay),_this select 0)]);
+    class ACE_RscHUDHelper {
         class controls {
-            class speedIndicator: RscPictureKeepAspect {
-                onLoad = QUOTE(uiNamespace setVariable [ARR_2(QQGVAR(speedIndicator),_this select 0)]);
-                idc = 10;
+            class GVAR(speedIndicator): RscPictureKeepAspect {
+                onLoad = QUOTE(_this call EFUNC(common,registerHUDElement));
+                condition = QUOTE(GVAR(enableSpeedIndicator) && {(vehicle ACE_player) == ACE_player});
+                idc = -1;
                 x = IGUI_GRID_STANCE_X + IGUI_GRID_STANCE_WAbs / 2 - SPEED_INDICATOR_W / 2;
                 y = IGUI_GRID_STANCE_Y + IGUI_GRID_STANCE_HAbs - SPEED_INDICATOR_Y_OFFSET;
                 w = SPEED_INDICATOR_W;
